app/requirements.txt
flask
python-dotenv
torch
openai-whisper
requests
google-genai
